Move aoi files into same folder
