+++ app/scripts/components/common/map/controls/aoi/types.d.ts
