+++ app/scripts/components/common/map/controls/aoi/utils.ts
